Move url_ext-specific errors to url_ext.rs
