+++ payjoin/src/uri/error.rs
@@ -11,86 +11,6 @@ pub(crate) enum InternalPjParseError {
     UnsecureEndpoint,
 }
 
-#[cfg(feature = "v2")]
-#[derive(Debug)]
-pub(crate) enum ParseOhttpKeysParamError {
-    MissingOhttpKeys,
-    InvalidOhttpKeys(crate::ohttp::ParseOhttpKeysError),
-}
-
-#[cfg(feature = "v2")]
-#[derive(Debug)]
-pub(crate) enum ParseExpParamError {
-    MissingExp,
-    InvalidHrp(bitcoin::bech32::Hrp),
-    DecodeBech32(bitcoin::bech32::primitives::decode::CheckedHrpstringError),
-    InvalidExp(bitcoin::consensus::encode::Error),
-}
-
-#[cfg(feature = "v2")]
-impl std::fmt::Display for ParseExpParamError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        use ParseExpParamError::*;
-
-        match &self {
-            MissingExp => write!(f, "exp is missing"),
-            InvalidHrp(h) => write!(f, "incorrect hrp for exp: {}", h),
-            DecodeBech32(d) => write!(f, "exp is not valid bech32: {}", d),
-            InvalidExp(i) => write!(f, "exp param does not contain a bitcoin consensus encoded u32: {}", i),
-        }
-    }
-}
-
-#[cfg(feature = "v2")]
-impl std::fmt::Display for ParseOhttpKeysParamError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        use ParseOhttpKeysParamError::*;
-
-        match &self {
-            MissingOhttpKeys => write!(f, "ohttp keys are missing"),
-            InvalidOhttpKeys(o) => write!(f, "invalid ohttp keys: {}", o),
-        }
-    }
-}
-
-#[cfg(feature = "v2")]
-#[derive(Debug)]
-pub(crate) enum ParseReceiverPubkeyParamError {
-    MissingPubkey,
-    InvalidHrp(bitcoin::bech32::Hrp),
-    DecodeBech32(bitcoin::bech32::primitives::decode::CheckedHrpstringError),
-    InvalidPubkey(crate::hpke::HpkeError),
-}
-
-#[cfg(feature = "v2")]
-impl std::fmt::Display for ParseReceiverPubkeyParamError {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        use ParseReceiverPubkeyParamError::*;
-
-        match &self {
-            MissingPubkey => write!(f, "receiver public key is missing"),
-            InvalidHrp(h) => write!(f, "incorrect hrp for receiver key: {}", h),
-            DecodeBech32(e) => write!(f, "receiver public is not valid base64: {}", e),
-            InvalidPubkey(e) =>
-                write!(f, "receiver public key does not represent a valid pubkey: {}", e),
-        }
-    }
-}
-
-#[cfg(feature = "v2")]
-impl std::error::Error for ParseReceiverPubkeyParamError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        use ParseReceiverPubkeyParamError::*;
-
-        match &self {
-            MissingPubkey => None,
-            InvalidHrp(_) => None,
-            DecodeBech32(error) => Some(error),
-            InvalidPubkey(error) => Some(error),
-        }
-    }
-}
-
 impl From<InternalPjParseError> for PjParseError {
     fn from(value: InternalPjParseError) -> Self { PjParseError(value) }
 }
